[GTK] Pass available input devices from UI process to web process for Interaction Media Features implementation https://bugs.webkit.org/show_bug.cgi?id=291192
Reviewed by Adrian Perez de Castro.

Using the same code added in 293276@main for WPE, but getting the devices using GTK API.

* Source/WebCore/platform/PlatformScreen.h:
(WebCore::screenHasTouchDevice):
(WebCore::screenIsTouchPrimaryInputDevice): Deleted.
* Source/WebCore/platform/gtk/PlatformScreenGtk.cpp:
(WebCore::screenHasTouchDevice):
(WebCore::screenIsTouchPrimaryInputDevice): Deleted.
* Source/WebKit/PlatformGTK.cmake:
* Source/WebKit/Scripts/webkit/messages.py:
(conditions_for_header):
* Source/WebKit/Shared/WebProcessCreationParameters.h:
* Source/WebKit/Shared/WebProcessCreationParameters.serialization.in:
* Source/WebKit/Shared/glib/AvailableInputDevices.h:
* Source/WebKit/Shared/glib/AvailableInputDevices.serialization.in:
* Source/WebKit/UIProcess/glib/WebProcessPoolGLib.cpp:
(WebKit::toAvailableInputDevices):
(WebKit::availableInputDevices):
(WebKit::seatDevicesChangedCallback):
(WebKit::WebProcessPool::platformInitialize):
(WebKit::WebProcessPool::platformInitializeWebProcess):
(WebKit::WebProcessPool::platformInvalidateContext):
* Source/WebKit/WebProcess/WebPage/glib/WebPageGLib.cpp:
(WebKit::WebPage::hoverSupportedByPrimaryPointingDevice const):
(WebKit::WebPage::hoverSupportedByAnyAvailablePointingDevice const):
(WebKit::WebPage::pointerCharacteristicsOfPrimaryPointingDevice const):
(WebKit::WebPage::pointerCharacteristicsOfAllAvailablePointingDevices const):
* Source/WebKit/WebProcess/WebPage/gtk/WebPageGtk.cpp:
(WebKit::WebPage::hoverSupportedByPrimaryPointingDevice const): Deleted.
(WebKit::WebPage::hoverSupportedByAnyAvailablePointingDevice const): Deleted.
(WebKit::WebPage::pointerCharacteristicsOfPrimaryPointingDevice const): Deleted.
(WebKit::WebPage::pointerCharacteristicsOfAllAvailablePointingDevices const): Deleted.
* Source/WebKit/WebProcess/WebPage/wpe/WebPageWPE.cpp:
(WebKit::WebPage::hoverSupportedByPrimaryPointingDevice const): Deleted.
(WebKit::WebPage::hoverSupportedByAnyAvailablePointingDevice const): Deleted.
(WebKit::WebPage::pointerCharacteristicsOfPrimaryPointingDevice const): Deleted.
(WebKit::WebPage::pointerCharacteristicsOfAllAvailablePointingDevices const): Deleted.
* Source/WebKit/WebProcess/WebProcess.h:
* Source/WebKit/WebProcess/WebProcess.messages.in:
* Source/WebKit/WebProcess/glib/WebProcessGLib.cpp:
(WebKit::WebProcess::platformInitializeWebProcess):
(WebKit::WebProcess::setScreenProperties):

Canonical link: https://commits.webkit.org/293352@main
